requirements.txt
pandas
numpy
boto3
pymongo
astrapy
python-dotenv
matplotlib